Support using integers as booleans in expressions
For MySQL compatibility, support using integers as booleans in
expressions.  This is an expensive and unreliable check, so limit it to
the cases currently observed in the wild.  We can expand the checks
later if more uses appear.

The current appearance is from the query:

SELECT SQL_CALC_FOUND_ROWS  wp_posts.ID FROM wp_posts  WHERE 1=1 AND 0
ORDER BY wp_posts.post_date DESC LIMIT 0, 10

made by require('wp-blog-header.php'), wp, WP->main, WP->query_posts,
WP_Query->query, WP_Query->get_posts

I can't determine the exact source of the calls or whether it is
URL-dependent.  But for me that is irrelevant, since it is a case that I
need to support.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>